fix: fix "makeRoutes" function
Details:

Accumulator in reduce function of Array.prototype was not used correctly and didn't accumulate
changes to routes object. So 'pizza/:pizzaId/toppings/:toppingId' in the end didn't become for example 'pizza/1/toppings/2'. Instead it became 'pizza/:pizzaId/toppings/2' which lead to app crash as the pathname wasn't correct.

We MUST use accumulator in order to accumulate values of each iteration!